Add git-head check when finish the feature

+++ scripts/services/Feature.ts
@@ -1,33 +1,30 @@
-import * as fs from "fs-extra";
 import {Git} from "./Git";
 import {Settings} from "./Settings";
 
 export class Feature {
     static start(destFolder: string, name: string) {
-        let featureFile: string = destFolder + "/feature.lock";
         let git: Git = new Git(Settings.folder);
         git.checkout(destFolder, "master");
         git.pull(destFolder);
 
-        if (!fs.existsSync(featureFile)) {
-            fs.createFileSync(featureFile);
-        }
-        fs.appendFileSync(featureFile, `[FEATURE] ${name} - Start at ${new Date().toDateString()} \n`);
-
         git.createBranch(destFolder, name);
-        git.add(destFolder, "feature.lock");
-        git.commit(destFolder, `Start feature ${name}`);
-        git.push(destFolder, "origin", name);
     }
 
     static finish(destFolder: string, name: string) {
         let git: Git = new Git(Settings.folder);
+        let storyId: string = name.split("-")[1];
+
         git.checkout(destFolder, name);
         git.pull(destFolder);
 
-        git.checkout(destFolder, "master");
-        git.merge(destFolder, name);
-        git.push(destFolder, "origin", "master");
-        git.deleteBranch(destFolder, name);
+        let headTag: string = git.headTag(destFolder);
+        if (headTag && headTag.split(".")[1] === storyId) {
+            git.checkout(destFolder, "master");
+            git.merge(destFolder, name);
+            git.push(destFolder, "origin", "master");
+            git.deleteBranch(destFolder, name);
+        }
+        else
+            console.error("[ERROR] Head not contains a valid tag");
     }
 }
